convert: size/capacity from int to long

+++ src/test/java/com/vetrikos/inventory/system/service/WarehouseServiceImplIT.java
@@ -1,8 +1,5 @@
 package com.vetrikos.inventory.system.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 import com.vetrikos.inventory.system.config.CustomPostgreSQLContainer;
 import com.vetrikos.inventory.system.config.IntegrationTest;
 import com.vetrikos.inventory.system.config.InventoryKeycloakContainer;
@@ -14,8 +11,6 @@ import com.vetrikos.inventory.system.model.WarehouseRequestRestDTO;
 import com.vetrikos.inventory.system.model.WarehouseUpdateRequestRestDTO;
 import com.vetrikos.inventory.system.repository.UserRepository;
 import com.vetrikos.inventory.system.repository.WarehouseRepository;
-import java.util.List;
-import java.util.UUID;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -23,166 +18,172 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.testcontainers.junit.jupiter.Container;
 
+import java.util.List;
+import java.util.UUID;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 @IntegrationTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 class WarehouseServiceImplIT {
 
-  @Container
-  public static CustomPostgreSQLContainer postgreSQLContainer = CustomPostgreSQLContainer.getInstance();
+    @Container
+    public static CustomPostgreSQLContainer postgreSQLContainer = CustomPostgreSQLContainer.getInstance();
 
-  @Container
-  public static InventoryKeycloakContainer keycloakContainer = InventoryKeycloakContainer.getInstance();
+    @Container
+    public static InventoryKeycloakContainer keycloakContainer = InventoryKeycloakContainer.getInstance();
 
-  @Autowired
-  private UserRepository userRepository;
+    @Autowired
+    private UserRepository userRepository;
 
-  @Autowired
-  private WarehouseRepository warehouseRepository;
+    @Autowired
+    private WarehouseRepository warehouseRepository;
 
-  @Autowired
-  private WarehouseServiceImpl warehouseService;
+    @Autowired
+    private WarehouseServiceImpl warehouseService;
 
-  @Autowired
-  private WarehouseMapper warehouseMapper;
+    @Autowired
+    private WarehouseMapper warehouseMapper;
 
-  private User sampleUser;
+    private User sampleUser;
 
-  @BeforeEach
-  void setUp() {
-    sampleUser = User.builder()
-        .id(UUID.randomUUID())
-        .username("user")
-        .fullName("Sample User")
-        .build();
-  }
+    @BeforeEach
+    void setUp() {
+        sampleUser = User.builder()
+                .id(UUID.randomUUID())
+                .username("user")
+                .fullName("Sample User")
+                .build();
+    }
 
-  @AfterEach
-  void tearDown() {
-    userRepository.deleteAll();
-    warehouseRepository.deleteAll();
-  }
+    @AfterEach
+    void tearDown() {
+        userRepository.deleteAll();
+        warehouseRepository.deleteAll();
+    }
 
-  @Test
-  void shouldFindAll() {
-    Warehouse warehouse = Warehouse.builder()
-        .capacity(100)
-        .name("Warehouse")
-        .users(List.of(sampleUser))
-        .build();
+    @Test
+    void shouldFindAll() {
+        Warehouse warehouse = Warehouse.builder()
+                .capacity(100L)
+                .name("Warehouse")
+                .users(List.of(sampleUser))
+                .build();
 
-    sampleUser.setWarehouse(warehouse);
-    sampleUser = userRepository.save(sampleUser);
+        sampleUser.setWarehouse(warehouse);
+        sampleUser = userRepository.save(sampleUser);
 
-    List<BasicWarehouse> result = warehouseService.findAll();
+        List<BasicWarehouse> result = warehouseService.findAll();
 
-    System.out.println(result);
-    System.out.println(warehouse.toString());
+        System.out.println(result);
+        System.out.println(warehouse.toString());
 
-    assertThat(result).isNotEmpty()
-        .hasSize(1);
-    assertThat(warehouseMapper.basicWarehouseToBasicWarehouseRestDTO(result.get(0))).isEqualTo(
-        warehouseMapper.warehouseToBasicWarehouseRestDTO(warehouse));
-  }
+        assertThat(result).isNotEmpty()
+                .hasSize(1);
+        assertThat(warehouseMapper.basicWarehouseToBasicWarehouseRestDTO(result.get(0))).isEqualTo(
+                warehouseMapper.warehouseToBasicWarehouseRestDTO(warehouse));
+    }
 
-  @Test
-  void shouldFindById() {
-    Warehouse warehouse = Warehouse.builder()
-        .capacity(100)
-        .name("Warehouse")
-        .users(List.of(sampleUser))
-        .build();
+    @Test
+    void shouldFindById() {
+        Warehouse warehouse = Warehouse.builder()
+                .capacity(100L)
+                .name("Warehouse")
+                .users(List.of(sampleUser))
+                .build();
 
-    sampleUser.setWarehouse(warehouse);
-    sampleUser = userRepository.save(sampleUser);
+        sampleUser.setWarehouse(warehouse);
+        sampleUser = userRepository.save(sampleUser);
 
-    Warehouse result = warehouseService.findById(warehouse.getId());
+        Warehouse result = warehouseService.findById(warehouse.getId());
 
-    assertThat(result).isNotNull()
-        .isEqualTo(warehouse);
-  }
+        assertThat(result).isNotNull()
+                .isEqualTo(warehouse);
+    }
 
-  @Test
-  void findByIdShouldThrowException() {
-    Long warehouseId = 1L;
-    assertThatThrownBy(() -> warehouseService.findById(warehouseId))
-        .hasMessage(WarehouseService.warehouseNotFoundMessage(warehouseId));
-  }
+    @Test
+    void findByIdShouldThrowException() {
+        Long warehouseId = 1L;
+        assertThatThrownBy(() -> warehouseService.findById(warehouseId))
+                .hasMessage(WarehouseService.warehouseNotFoundMessage(warehouseId));
+    }
 
-  @Test
-  void shouldCreateWarehouse() {
-    sampleUser = userRepository.save(sampleUser);
+    @Test
+    void shouldCreateWarehouse() {
+        sampleUser = userRepository.save(sampleUser);
 
-    int capacity = 100;
-    String warehouseName = "Lidl warehouse";
-    WarehouseRequestRestDTO requestRestDTO = new WarehouseRequestRestDTO(capacity, warehouseName);
+        long capacity = 100;
+        String warehouseName = "Lidl warehouse";
+        WarehouseRequestRestDTO requestRestDTO = new WarehouseRequestRestDTO(capacity, warehouseName);
 
-    Warehouse result = warehouseService.createWarehouse(requestRestDTO);
+        Warehouse result = warehouseService.createWarehouse(requestRestDTO);
 
-    Warehouse expectedWarehouse = Warehouse.builder()
-        .id(result.getId())
-        .capacity(capacity)
-        .name(warehouseName)
+        Warehouse expectedWarehouse = Warehouse.builder()
+                .id(result.getId())
+                .capacity(capacity)
+                .name(warehouseName)
 //        .users(List.of(sampleUser))
-        .build();
+                .build();
 
-    assertThat(result).isNotNull()
-        .isEqualTo(expectedWarehouse);
-    assertThat(warehouseRepository.findAll().size()).isOne();
-  }
+        assertThat(result).isNotNull()
+                .isEqualTo(expectedWarehouse);
+        assertThat(warehouseRepository.findAll().size()).isOne();
+    }
 
-  @Test
-  void shouldUpdateWarehouse() {
-    Warehouse warehouse = Warehouse.builder()
-        .capacity(100)
-        .name("Warehouse")
-        .users(List.of(sampleUser))
-        .build();
+    @Test
+    void shouldUpdateWarehouse() {
+        Warehouse warehouse = Warehouse.builder()
+                .capacity(100L)
+                .name("Warehouse")
+                .users(List.of(sampleUser))
+                .build();
 
-    sampleUser.setWarehouse(warehouse);
-    sampleUser = userRepository.save(sampleUser);
+        sampleUser.setWarehouse(warehouse);
+        sampleUser = userRepository.save(sampleUser);
 
-    Long warehouseId = warehouse.getId();
-    int newCapacity = 200;
-    String newWarehouseName = "Lidl warehouse";
-    WarehouseUpdateRequestRestDTO requestRestDTO = new WarehouseUpdateRequestRestDTO();
-    requestRestDTO.setCapacity(newCapacity);
-    requestRestDTO.setName(newWarehouseName);
+        Long warehouseId = warehouse.getId();
+        long newCapacity = 200;
+        String newWarehouseName = "Lidl warehouse";
+        WarehouseUpdateRequestRestDTO requestRestDTO = new WarehouseUpdateRequestRestDTO();
+        requestRestDTO.setCapacity(newCapacity);
+        requestRestDTO.setName(newWarehouseName);
 
-    Warehouse result = warehouseService.updateWarehouse(warehouseId, requestRestDTO);
+        Warehouse result = warehouseService.updateWarehouse(warehouseId, requestRestDTO);
 
-    assertThat(result).isNotNull()
-        .hasFieldOrPropertyWithValue("capacity", newCapacity)
-        .hasFieldOrPropertyWithValue("name", newWarehouseName);
-  }
+        assertThat(result).isNotNull()
+                .hasFieldOrPropertyWithValue("capacity", newCapacity)
+                .hasFieldOrPropertyWithValue("name", newWarehouseName);
+    }
 
-  @Test
-  void updateWarehouseShouldThrowException() {
-    Long warehouseId = 1L;
-    WarehouseUpdateRequestRestDTO requestRestDTO = new WarehouseUpdateRequestRestDTO();
-    assertThatThrownBy(() -> warehouseService.updateWarehouse(warehouseId, requestRestDTO))
-        .hasMessage(WarehouseService.warehouseNotFoundMessage(warehouseId));
-  }
+    @Test
+    void updateWarehouseShouldThrowException() {
+        Long warehouseId = 1L;
+        WarehouseUpdateRequestRestDTO requestRestDTO = new WarehouseUpdateRequestRestDTO();
+        assertThatThrownBy(() -> warehouseService.updateWarehouse(warehouseId, requestRestDTO))
+                .hasMessage(WarehouseService.warehouseNotFoundMessage(warehouseId));
+    }
 
-  @Test
-  void shouldDeleteWarehouse() {
-    Warehouse warehouse = Warehouse.builder()
-        .capacity(100)
-        .name("Warehouse")
-        .users(List.of(sampleUser))
-        .build();
+    @Test
+    void shouldDeleteWarehouse() {
+        Warehouse warehouse = Warehouse.builder()
+                .capacity(100L)
+                .name("Warehouse")
+                .users(List.of(sampleUser))
+                .build();
 
-    sampleUser.setWarehouse(warehouse);
-    sampleUser = userRepository.save(sampleUser);
+        sampleUser.setWarehouse(warehouse);
+        sampleUser = userRepository.save(sampleUser);
 
-    warehouseService.deleteWarehouse(warehouse.getId());
+        warehouseService.deleteWarehouse(warehouse.getId());
 
-    assertThat(warehouseRepository.findAll()).isEmpty();
-    assertThat(userRepository.findAll()).contains(sampleUser);
-  }
+        assertThat(warehouseRepository.findAll()).isEmpty();
+        assertThat(userRepository.findAll()).contains(sampleUser);
+    }
 
-  @Test
-  void deleteWarehouseShouldThrowException() {
-    Long warehouseId = 1L;
-    assertThatThrownBy(() -> warehouseService.deleteWarehouse(warehouseId))
-        .hasMessage(WarehouseService.warehouseNotFoundMessage(warehouseId));
-  }
+    @Test
+    void deleteWarehouseShouldThrowException() {
+        Long warehouseId = 1L;
+        assertThatThrownBy(() -> warehouseService.deleteWarehouse(warehouseId))
+                .hasMessage(WarehouseService.warehouseNotFoundMessage(warehouseId));
+    }
 }
